requirements.txt
# ==========================================
# FaceForge AI - ZeroGPU Gradio Requirements
# ==========================================

# Gradio & ZeroGPU (REQUIRED)
gradio==4.16.0
spaces==0.19.4

# Image Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3

# Background Removal
rembg[gpu]==2.0.57

# Face Enhancement (GPU)
gfpgan==1.3.8
realesrgan==0.3.0
basicsr==1.4.2

# PyTorch (GPU) - Critical for CUDA
torch==2.1.2
torchvision==0.16.2
#--extra-index-url https://download.pytorch.org/whl/cu118
--extra-index-url https://download.pytorch.org/whl/cu121

# Stable Diffusion 
diffusers>=0.20.0


# Core Dependencies
scipy==1.12.0
tqdm==4.66.1
facexlib==0.3.0
yapf==0.40.2
filterpy==1.4.5

# Optional: For better performance
# onnxruntime-gpu==1.17.0